pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Ammaro173 <ammaraf97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
